# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Documentation",
        "description": "Store API Postman and Code definitions documentation.",
        "contact": {
            "name": "Backend Engineer"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List Products",
                "description": "Get a paginated list of published products with optional filters",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price_min",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "in_stock",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{slug}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Product details",
                "description": "Get full details of a published product by its slug",
                "operationId": "1752983014e15d384cd3812cdff6367f",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            }
        },
        "/api/v1/products/{id}/reviews": {
            "get": {
                "tags": [
                    "Products",
                    "Reviews"
                ],
                "summary": "Get Product Reviews",
                "description": "Get a paginated list of approved reviews for a specific product",
                "operationId": "2a56b515bfb464ab80b7f8892b48ee4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Review"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotFoundResponse": {
                "title": "NotFound Response",
                "description": "Response for 404 not found error",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "Pagination Meta",
                "description": "Metadata for paginated responses (Laravel style)",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "example": "http://example.com/api?page=1",
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string",
                                    "example": "1"
                                },
                                "active": {
                                    "type": "boolean",
                                    "example": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "path": {
                        "type": "string",
                        "example": "http://example.com/api"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 75
                    }
                },
                "type": "object"
            },
...
            "CategoryBasic": {
                "title": "CategoryBasic",
                "description": "Minimal category reference (used inside product listings)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "slug": {
                        "type": "string",
                        "example": "electronics"
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "title": "Product",
                "description": "Product schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "MacBook Pro"
                    },
                    "slug": {
                        "type": "string",
                        "example": "macbook-pro"
                    },
                    "description": {
                        "type": "string",
                        "example": "High performance laptop",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1999.99
                    },
                    "discount_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1799.99,
                        "nullable": true
                    },
                    "final_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1799.99
                    },
                    "stock_qty": {
                        "type": "integer",
                        "example": 10
                    },
                    "sku": {
                        "type": "string",
                        "example": "APL-MBC-2023",
                        "nullable": true
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": false
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.8
                    },
                    "reviews_count": {
                        "type": "integer",
                        "example": 15
                    },
                    "category": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CategoryBasic"
                            }
                        ],
                        "nullable": true
                    },
                    "image": {
                        "properties": {
                            "thumb": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "card": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "detail": {
                                "description": "Only in product detail endpoint",
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "zoom": {
                                "description": "Only in product detail endpoint",
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "gallery": {
                        "description": "Only in product detail endpoint",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "thumb": {
                                    "type": "string",
                                    "format": "uri",
                                    "nullable": true
                                },
                                "card": {
                                    "type": "string",
                                    "format": "uri",
                                    "nullable": true
                                },
                                "detail": {
                                    "type": "string",
                                    "format": "uri",
                                    "nullable": true
                                },
                                "zoom": {
                                    "type": "string",
                                    "format": "uri",
                                    "nullable": true
                                },
                                "order": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "variants": {
                        "description": "Only in product detail endpoint",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariant"
                        },
                        "nullable": true
                    },
                    "attributes": {
                        "description": "Only in product detail endpoint",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductVariant": {
                "title": "ProductVariant",
                "description": "Product variant (color/size combination)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Red / XL"
                    },
                    "sku": {
                        "type": "string",
                        "example": "SHIRT-RED-XL",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 35,
                        "nullable": true
                    },
                    "stock_qty": {
                        "type": "integer",
                        "example": 50
                    },
                    "attributes": {
                        "type": "object",
                        "example": {
                            "Color": "Red",
                            "Size": "XL"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Review": {
                "title": "Review",
                "description": "Product Review schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "rating": {
                        "type": "integer",
                        "example": 5
                    },
                    "comment": {
                        "type": "string",
                        "example": "Great product!",
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "name": {
                                    "type": "string",
                                    "example": "User"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "NotFoundResponse": {
                "description": "Not found error response",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFoundResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}